refactor(billing): split 1607-line billing page into focused modules
Decompose app/billing/page.tsx (1607 lines) into 11 files, all well under
the 400-500 line target. Purely structural — no behavior change.

- types.ts: shared interfaces + AxiosInstance alias
- utils.ts: fmt/round2/errMsg/labelFor/statusColors/timeAgo/orderLabel/…
- components/: PaymentPanel, OrderDetailPanel, OrderList, HistoryView,
  ReceiptHistoryDetail, ReceiptPrintArea, ReceiptHistoryPrintArea,
  PaymentMethodButton, BalanceSummary
- page.tsx is now a 295-line orchestrator (fetch, socket sync, tabs, layout)

Colocated under app/billing/ so nothing else in the app is affected.
tsc --noEmit passes; next build compiles /billing cleanly.
